Add pause menu. Add font. Add text component.

+++ app/src/main/java/io/github/spaceshooter/engine/GameView.java
@@ -32,8 +32,8 @@ public interface GameView {
         return getHeight() - getPaddingTop() - getPaddingBottom();
     }
 
-    default int getGUIWidth() {
-        return getFinalWidth() / getFinalHeight();
+    default float getGUIWidth() {
+        return getFinalWidth() / (float) getFinalHeight();
     }
 
     default Vector2f screenToGUI(Vector2f vec) {
